+++ frontendproject/src/components/sidebar.vue
@@ -34,6 +34,7 @@
 
 <script>
 import axios from "axios";
+import swal from 'sweetalert';
 
 export default {
 name: 'sidebarComponent',
@@ -48,9 +49,27 @@ data() {
   },
    methods: {
     logout() {
-      localStorage.clear();
+      swal({
+  title: "Are you sure?",
+  text: " you will Logout!",
+  icon: "warning",
+  buttons: true,
+  dangerMode: true,
+})
+.then((willDelete) => {
+  if (willDelete) {
+    localStorage.clear();
       axios.post("http://127.0.0.1:8000/api/logout");
-      this.$router.push("/login");
+      
+    swal("you are loged out!", {
+      icon: "success",
+    });
+    this.$router.push("/login");
+  } else {
+    swal("You are not loged out");
+  }
+});
+      
     },
   },
 }
